Move SystemsComplianceFilter to PresentationalComponents
