+++ src/PresentationalComponents/index.js
@@ -3,4 +3,5 @@ export { default as LoadingComplianceCards } from './LoadingComplianceCards/Load
 export { default as CompliancePolicyCard } from './CompliancePolicyCard/CompliancePolicyCard';
 export { default as CompliancePoliciesEmptyState } from './CompliancePoliciesEmptyState/CompliancePoliciesEmptyState';
 export { default as ErrorPage } from './ErrorPage/ErrorPage';
+export { default as SystemsComplianceFilter } from './SystemsComplianceFilter/SystemsComplianceFilter';
 export { default as PolicyDetailsContentLoader } from './PolicyDetailsContentLoader/PolicyDetailsContentLoader';
