+++ src/SmartComponents/index.js
@@ -1,0 +1,2 @@
+export { default as ComplianceRemediationButton } from './ComplianceRemediationButton/ComplianceRemediationButton';
+export { default as DownloadTableButton } from './DownloadTableButton/DownloadTableButton';
